+++ Demo/src/main/java/com/braintreepayments/demo/PayPalRequestFactory.java
@@ -13,6 +13,7 @@ import com.braintreepayments.api.paypal.PayPalPaymentUserAction;
 import com.braintreepayments.api.paypal.PayPalPricingModel;
 import com.braintreepayments.api.paypal.PayPalRecurringBillingDetails;
 import com.braintreepayments.api.paypal.PayPalRecurringBillingPlanType;
+import com.braintreepayments.api.paypal.PayPalPhoneNumber;
 import com.braintreepayments.api.paypal.PayPalVaultRequest;
 
 import java.util.ArrayList;
@@ -22,13 +23,19 @@ public class PayPalRequestFactory {
 
     public static PayPalVaultRequest createPayPalVaultRequest(
         Context context,
-        String buyerEmailAddress
+        String buyerEmailAddress,
+        String buyerPhoneCountryCode,
+        String buyerPhoneNationalNumber
     ) {
 
         PayPalVaultRequest request = new PayPalVaultRequest(true);
 
         if (!buyerEmailAddress.isEmpty()) {
             request.setUserAuthenticationEmail(buyerEmailAddress);
+        }
+
+        if (!buyerPhoneCountryCode.isEmpty() && !buyerPhoneNationalNumber.isEmpty()) {
+            request.setUserPhoneNumber(new PayPalPhoneNumber(buyerPhoneCountryCode, buyerPhoneNationalNumber));
         }
 
         if (Settings.isPayPalAppSwithEnabled(context)) {
@@ -102,12 +109,18 @@ public class PayPalRequestFactory {
     public static PayPalCheckoutRequest createPayPalCheckoutRequest(
         Context context,
         String amount,
-        String buyerEmailAddress
+        String buyerEmailAddress,
+        String buyerPhoneCountryCode,
+        String buyerPhoneNationalNumber
     ) {
         PayPalCheckoutRequest request = new PayPalCheckoutRequest(amount, true);
 
         if (!buyerEmailAddress.isEmpty()) {
             request.setUserAuthenticationEmail(buyerEmailAddress);
+        }
+
+        if (!buyerPhoneCountryCode.isEmpty() && !buyerPhoneNationalNumber.isEmpty()) {
+            request.setUserPhoneNumber(new PayPalPhoneNumber(buyerPhoneCountryCode, buyerPhoneNationalNumber));
         }
 
         request.setDisplayName(Settings.getPayPalDisplayName(context));