+++ Demo/src/main/java/com/braintreepayments/demo/ShopperInsightsFragment.kt
@@ -220,7 +220,9 @@ class ShopperInsightsFragment : BaseFragment() {
             requireContext(),
             PayPalRequestFactory.createPayPalVaultRequest(
                 activity,
-                emailInput.editText?.text.toString()
+                emailInput.editText?.text.toString(),
+                countryCodeInput.editText?.text.toString(),
+                nationalNumberInput.editText?.text.toString()
             )
         ) { authRequest ->
             when (authRequest) {