+++ PayPal/src/main/java/com/braintreepayments/api/paypal/PayPalRequest.kt
@@ -70,6 +70,8 @@ import org.json.JSONException
  * server.
  * @property userAuthenticationEmail User email to initiate a quicker authentication flow in cases
  * where the user has a PayPal Account with the same email.
+ * @property userPhoneNumber User phone number used to initiate a quicker authentication flow in
+ * cases where the user has a PayPal Account with the phone number.
  * @property lineItems The line items for this transaction. It can include up to 249 line items.
  */
 abstract class PayPalRequest internal constructor(
@@ -84,6 +86,7 @@ abstract class PayPalRequest internal constructor(
     open var merchantAccountId: String? = null,
     open var riskCorrelationId: String? = null,
     open var userAuthenticationEmail: String? = null,
+    open var userPhoneNumber: PayPalPhoneNumber? = null,
     open var lineItems: List<PayPalLineItem> = emptyList()
 ) : Parcelable {
 
@@ -128,5 +131,6 @@ abstract class PayPalRequest internal constructor(
         internal const val MERCHANT_APP_RETURN_URL_KEY: String = "merchant_app_return_url"
         internal const val PLAN_TYPE_KEY: String = "plan_type"
         internal const val PLAN_METADATA_KEY: String = "plan_metadata"
+        internal const val PHONE_NUMBER_KEY: String = "phone_number"
     }
 }